# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Claim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "kind": {
            "enum": [
              "observation",
              "interpretation",
              "recommendation"
            ],
            "title": "Kind",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "field",
          "text",
          "kind",
          "evidence_ids"
        ],
        "title": "Claim",
        "type": "object"
      },
...
      "DataIssue": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "line_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Line Ids",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "rows": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Rows",
            "type": "array"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "status": {
            "default": "needs_data",
            "enum": [
              "needs_review",
              "needs_data",
              "unresolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "action"
        ],
        "title": "DataIssue",
        "type": "object"
      },
      "Entity": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "type": {
            "const": "product",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "display_name"
        ],
        "title": "Entity",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "errors": {
            "default": [],
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/DataIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "issues"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvidenceMeta": {
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "locator": {
            "additionalProperties": true,
            "title": "Locator",
            "type": "object"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "sales_csv",
              "invoice_pdf",
              "calculation",
              "calendar"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "source_id",
          "source_type",
          "label",
          "locator"
        ],
        "title": "EvidenceMeta",
        "type": "object"
      },
      "EvidenceResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "locator": {
            "additionalProperties": true,
            "title": "Locator",
            "type": "object"
          },
          "page_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Image Url"
          },
          "signal_id": {
            "title": "Signal Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "sales_csv",
              "invoice_pdf",
              "calculation",
              "calendar"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "signal_id",
          "source_id",
          "source_type",
          "label",
          "locator",
          "excerpt"
        ],
        "title": "EvidenceResponse",
        "type": "object"
      },
      "FinancialImpact": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "currency": {
            "const": "INR",
            "default": "INR",
            "title": "Currency",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "label",
          "method"
        ],
        "title": "FinancialImpact",
        "type": "object"
      },
...
      "ModelMetadata": {
        "additionalProperties": false,
        "properties": {
          "detector_version": {
            "title": "Detector Version",
            "type": "string"
          },
          "explanation_source": {
            "enum": [
              "llm",
              "template"
            ],
            "title": "Explanation Source",
            "type": "string"
          },
          "extraction_method": {
            "title": "Extraction Method",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          }
        },
        "required": [
          "detector_version",
          "prompt_version",
          "model_name",
          "explanation_source",
          "extraction_method"
        ],
        "title": "ModelMetadata",
        "type": "object"
      },
...
      "RejectedCandidate": {
        "additionalProperties": false,
        "properties": {
          "anomaly_score": {
            "title": "Anomaly Score",
            "type": "number"
          },
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "candidate_type": {
            "title": "Candidate Type",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "product_ids": {
            "items": {
              "type": "string"
            },
            "title": "Product Ids",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "candidate_type",
          "product_ids",
          "title",
          "reason",
          "anomaly_score",
          "evidence"
        ],
        "title": "RejectedCandidate",
        "type": "object"
      },
      "RejectedExplanation": {
        "additionalProperties": false,
        "properties": {
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "name",
          "reason",
          "evidence_ids"
        ],
        "title": "RejectedExplanation",
        "type": "object"
      },
      "Review": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "confirmed",
              "dismissed",
              "unresolved"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "signal_id": {
            "title": "Signal Id",
            "type": "string"
          }
        },
        "required": [
          "review_id",
          "signal_id",
          "run_id",
          "outcome",
          "reason",
          "created_at"
        ],
        "title": "Review",
        "type": "object"
      },
...
      "Signal": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "minItems": 1,
            "title": "Claims",
            "type": "array"
          },
          "entity": {
            "$ref": "#/components/schemas/Entity"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 2,
            "title": "Evidence Ids",
            "type": "array"
          },
          "evidence_strength": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Evidence Strength",
            "type": "number"
          },
          "evidence_strength_components": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Evidence Strength Components",
            "type": "object"
          },
          "facts": {
            "additionalProperties": true,
            "title": "Facts",
            "type": "object"
          },
          "financial_impact": {
            "$ref": "#/components/schemas/FinancialImpact"
          },
          "interpretation": {
            "title": "Interpretation",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Limitations",
            "type": "array"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata"
          },
          "next_check": {
            "title": "Next Check",
            "type": "string"
          },
          "observation": {
            "title": "Observation",
            "type": "string"
          },
          "ranking": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Ranking",
            "type": "object"
          },
          "rejected_explanations": {
            "items": {
              "$ref": "#/components/schemas/RejectedExplanation"
            },
            "title": "Rejected Explanations",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "severity": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Severity",
            "type": "string"
          },
          "signal_id": {
            "title": "Signal Id",
            "type": "string"
          },
          "signal_type": {
            "enum": [
              "margin_leakage",
              "inventory_discrepancy"
            ],
            "title": "Signal Type",
            "type": "string"
          },
          "status": {
            "default": "new",
            "enum": [
              "new",
              "unresolved",
              "confirmed",
              "dismissed",
              "needs_data"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "translations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SignalTranslation"
            },
            "title": "Translations",
            "type": "object"
          }
        },
        "required": [
          "signal_id",
          "run_id",
          "signal_type",
          "severity",
          "evidence_strength",
          "evidence_strength_components",
          "entity",
          "title",
          "observation",
          "financial_impact",
          "interpretation",
          "evidence_ids",
          "rejected_explanations",
          "limitations",
          "next_check",
          "facts",
          "ranking",
          "model_metadata",
          "claims"
        ],
        "title": "Signal",
        "type": "object"
      },
      "SignalDetail": {
        "properties": {
          "evidence": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EvidenceMeta"
            },
            "title": "Evidence",
            "type": "object"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/Review"
            },
            "title": "Reviews",
            "type": "array"
          },
          "signal": {
            "$ref": "#/components/schemas/Signal"
          }
        },
        "required": [
          "signal",
          "evidence",
          "reviews"
        ],
        "title": "SignalDetail",
        "type": "object"
      },
      "SignalSummary": {
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Entity"
          },
          "evidence_strength": {
            "title": "Evidence Strength",
            "type": "number"
          },
          "financial_impact": {
            "$ref": "#/components/schemas/FinancialImpact"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata"
          },
          "observation": {
            "title": "Observation",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "severity": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Severity",
            "type": "string"
          },
          "signal_id": {
            "title": "Signal Id",
            "type": "string"
          },
          "signal_type": {
            "enum": [
              "margin_leakage",
              "inventory_discrepancy"
            ],
            "title": "Signal Type",
            "type": "string"
          },
          "status": {
            "enum": [
              "new",
              "unresolved",
              "confirmed",
              "dismissed",
              "needs_data"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "translations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SignalTranslation"
            },
            "default": {},
            "title": "Translations",
            "type": "object"
          }
        },
        "required": [
          "signal_id",
          "run_id",
          "signal_type",
          "status",
          "severity",
          "evidence_strength",
          "entity",
          "title",
          "observation",
          "financial_impact",
          "model_metadata"
        ],
        "title": "SignalSummary",
        "type": "object"
      },
      "SignalTranslation": {
        "additionalProperties": false,
        "properties": {
          "impact_label": {
            "default": "",
            "title": "Impact Label",
            "type": "string"
          },
          "interpretation": {
            "default": "",
            "title": "Interpretation",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "next_check": {
            "default": "",
            "title": "Next Check",
            "type": "string"
          },
          "observation": {
            "default": "",
            "title": "Observation",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "SignalTranslation",
        "type": "object"
      },
      "SignalsResponse": {
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/DataIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "rejected_candidates": {
            "items": {
              "$ref": "#/components/schemas/RejectedCandidate"
            },
            "title": "Rejected Candidates",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "title": "Run Status",
            "type": "string"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalSummary"
            },
            "title": "Signals",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "run_status",
          "signals",
          "rejected_candidates",
          "issues"
        ],
        "title": "SignalsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Evidence-backed profit leakage signals for small grocery retailers. Investigation leads only; no automatic actions.",
    "title": "Tremor Retail API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/evidence/{evidence_id}": {
      "get": {
        "operationId": "get_evidence_api_evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Evidence"
      }
    },
...
    "/api/runs/{run_id}/signals": {
      "get": {
        "operationId": "list_signals_api_runs__run_id__signals_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Signals"
      }
    },
...
    "/api/signals/{signal_id}": {
      "get": {
        "operationId": "get_signal_api_signals__signal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "signal_id",
            "required": true,
            "schema": {
              "title": "Signal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Signal"
      }
    },
...
  }
}